git_ui: Remove dependency on agent crates
