+++ crates/git_ui/Cargo.toml
@@ -16,10 +16,7 @@ path = "src/git_ui.rs"
 test-support = ["multi_buffer/test-support", "remote_connection/test-support"]
 
 [dependencies]
-acp_thread.workspace = true
-agent-client-protocol.workspace = true
 agent_settings.workspace = true
-agent_ui.workspace = true
 anyhow.workspace = true
 askpass.workspace = true
 buffer_diff.workspace = true
